epoll related issue fix proposal

+++ api/model/src/main/java/com/emc/mongoose/api/model/data/SeedDataInput.java
@@ -17,8 +17,8 @@ extends CachedDataInput {
 		super();
 	}
 
-	public SeedDataInput(final long seed, final long size, final int cacheLimit) {
-		super((MappedByteBuffer) ByteBuffer.allocateDirect((int) size), cacheLimit);
+	public SeedDataInput(final long seed, final int size, final int cacheLimit) {
+		super((MappedByteBuffer) ByteBuffer.allocateDirect(size), cacheLimit);
 		generateData(inputBuff, seed);
 	}
